Re-work the handling of segments in the assembler
There were a lot of assumptions on the assembler that stemmed from a
fundamental missunderstanding from my side on how segments are laid out
on the final file. This commit is the first step to address this.

+++ lib/xixanta/src/mapping.rs
@@ -51,7 +51,7 @@ pub struct Segment {
     pub start: u16,
     pub size: usize,
     pub offset: usize,
-    pub fill: Option<usize>,
+    pub fill: Option<u8>,
     pub bundles: Vec<Bundle>,
 }
